Update README.md
Something had been wrong, for quite a while, with the Terminal. It was fine at first, but then it stopped showing. 

It was tough to fix and I'm not quite sure what the issue was (wouldn't work no matter if as MD or HTML). But looks like replacing the h1 with a Div did fix whatever was wrong with it.